release: load config at runtime, improve documentation
