+++ dev/release/release-config.jsonc
@@ -1,0 +1,40 @@
+/**
+ * Configuration for the release tool. All fields should align exactly with those in
+ * `src/config.ts`. To learn more about the release tool:
+ *
+ *     yarn run release help
+ *
+ * A quick sanity-check can be performed to make sure configuration is loaded correctly:
+ *
+ *     yarn run release _test:config
+ */
+{
+    // Captain information
+    "captainSlackUsername": "Gonza",
+    "captainGitHubUsername": "pecigonzalo",
+
+    // Release versions
+    "previousRelease": "3.21.2",
+    "upcomingRelease": "3.22.0",
+
+    // Release dates
+    "releaseDateTime": "20 November 2020 10:00 PST",
+    "oneWorkingDayBeforeRelease": "19 November 2020 10:00 PST",
+    "fourWorkingDaysBeforeRelease": "16 November 2020 10:00 PST",
+    "fiveWorkingDaysBeforeRelease": "13 November 2020 10:00 PST",
+
+    // Channel where messages from the tooling are posted
+    "slackAnnounceChannel": "dev-announce",
+
+    // Email for preparing calendar events
+    "teamEmail": "team@sourcegraph.com",
+
+    // Enable dry-running for certain features - useful for testing or sanity-checking.
+    //
+    // For example, `dryRun.changesets` prints changes generated by `release:stage`
+    // instead of pushing them to GitHub.
+    "dryRun": {
+        "changesets": false,
+        "trackingIssues": false
+    }
+}
