--- dev/release/src/config.json
@@ -1,21 +1,0 @@
-{
-  "teamEmail": "team@sourcegraph.com",
-
-  "captainSlackUsername": "Gonza",
-  "captainGitHubUsername": "pecigonzalo",
-
-  "previousRelease": "3.21.2",
-  "upcomingRelease": "3.22.0",
-
-  "releaseDateTime": "20 November 2020 10:00 PST",
-  "oneWorkingDayBeforeRelease": "19 November 2020 10:00 PST",
-  "fourWorkingDaysBeforeRelease": "16 November 2020 10:00 PST",
-  "fiveWorkingDaysBeforeRelease": "13 November 2020 10:00 PST",
-
-  "slackAnnounceChannel": "dev-announce",
-
-  "dryRun": {
-    "changesets": false,
-    "trackingIssues": false
-  }
-}
